split TETagQuery.java into multiple files
